requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
blis==0.2.4
boto==2.49.0
boto3==1.9.200
botocore==1.12.200
bz2file==0.98
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
gensim==3.7.3
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
murmurhash==1.0.0
nb-conda-kernels==2.2.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.2.5
nose==1.3.7
notebook==6.0.0
numpy==1.17.0
oauthlib==3.0.1
olefile==0.46
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
plyfile==0.7
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytz==2019.2
pyzmq==18.0.2
qtconsole==4.5.2
rdflib==4.2.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
smart-open==1.8.4
snorkel-metal==0.1.5
spacy==2.1.7
srsly==0.0.7
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
torch==1.2.0
torch-cluster==1.4.3
torch-geometric==1.3.0
torch-scatter==1.3.1
torchvision==0.4.0a0+9232c4a
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
twython==3.7.0
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
